Move scheduled inactive subject deletion to SubjectService, update tests
