+++ src/test/java/org/radarcns/management/service/SubjectServiceTest.java
@@ -1,23 +1,38 @@
 package org.radarcns.management.service;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.radarcns.management.service.dto.ProjectDTO.PRIVACY_POLICY_URL;
 import static org.radarcns.management.service.dto.SubjectDTO.SubjectStatus.ACTIVATED;
+import static org.radarcns.management.web.rest.TestUtil.commitTransactionAndStartNew;
 
 import java.net.URL;
+import java.time.Period;
+import java.time.ZonedDateTime;
 import java.util.Collections;
+import java.util.Date;
+import java.util.List;
 
+import org.hibernate.envers.AuditReader;
+import org.hibernate.envers.query.AuditEntity;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.radarcns.management.ManagementPortalTestApp;
 import org.radarcns.management.domain.Subject;
+import org.radarcns.management.domain.User;
+import org.radarcns.management.domain.audit.CustomRevisionEntity;
+import org.radarcns.management.repository.SubjectRepository;
+import org.radarcns.management.repository.UserRepository;
 import org.radarcns.management.service.dto.ProjectDTO;
 import org.radarcns.management.service.dto.SubjectDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.transaction.annotation.Transactional;
+
+import javax.persistence.EntityManager;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = ManagementPortalTestApp.class)
@@ -47,6 +62,15 @@ public class SubjectServiceTest {
 
     @Autowired
     private ProjectService projectService;
+
+    @Autowired
+    private SubjectRepository subjectRepository;
+
+    @Autowired
+    private RevisionService revisionService;
+
+    @Autowired
+    private UserRepository userRepository;
 
 
     /**
@@ -87,5 +111,45 @@ public class SubjectServiceTest {
         assertNotNull(privacyPolicyUrl);
         assertEquals(privacyPolicyUrl.toExternalForm(), DEFAULT_PROJECT_PRIVACY_POLICY_URL);
 
+    }
+
+    @Test
+    public void testFindNotActivatedSubjectsByCreationDateBefore() {
+        User expiredUser = UserServiceIntTest.addExpiredUser(userRepository);
+        Subject expiredSubject = new Subject();
+        expiredSubject.setUser(expiredUser);
+        subjectRepository.save(expiredSubject);
+        commitTransactionAndStartNew();
+
+        AuditReader auditReader = ((AuditReader) ReflectionTestUtils
+                .getField(revisionService, "auditReader"));
+        Object[] firstRevision = (Object[]) auditReader.createQuery()
+                .forRevisionsOfEntity(expiredUser.getClass(), false, true)
+                .add(AuditEntity.id().eq(expiredUser.getId()))
+                .add(AuditEntity.revisionNumber().minimize()
+                        .computeAggregationInInstanceContext())
+                .getSingleResult();
+        CustomRevisionEntity first = (CustomRevisionEntity) firstRevision[1];
+        // Update the timestamp of the revision so it appears to have been created 5 days ago
+        ZonedDateTime expDateTime = ZonedDateTime.now().minus(Period.ofDays(5));
+        first.setTimestamp(Date.from(expDateTime.toInstant()));
+        EntityManager entityManager = ((EntityManager) ReflectionTestUtils
+                .getField(revisionService, "entityManager"));
+        entityManager.persist(first);
+
+        // make sure when we reload the expired user we have the new created date
+        assertThat(revisionService.getAuditInfo(expiredUser).getCreatedAt()).isEqualTo(expDateTime);
+
+        // Now we know we have an 'old' user in the database, we can test our deletion method
+        int numSubjects = subjectRepository.findAll().size();
+        subjectService.removeNotActivatedSubjects();
+        List<Subject> subjects = subjectRepository.findAll();
+        // make sure have actually deleted some users, otherwise this test is pointless
+        assertThat(numSubjects - subjects.size()).isEqualTo(1);
+        // remaining users should be either activated or have a created date less then 3 days ago
+        ZonedDateTime cutoff = ZonedDateTime.now().minus(Period.ofDays(3));
+        subjects.forEach(s -> assertThat(s.getUser().getActivated() || revisionService
+                .getAuditInfo(s).getCreatedAt().isAfter(cutoff)).isTrue());
+        commitTransactionAndStartNew();
     }
 }
